video-creator/requirements.txt
psycopg2-binary==2.9.7
requests==2.31.0
moviepy==1.0.3
pillow==10.0.0
gtts==2.3.2
youtube-upload==0.8.0